Re #2209: Fixed build failure on Android due to incompatible type
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@6036 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/swig/java/android/app/src/main/java/org/pjsip/pjsua2/app/MyApp.java
@@ -28,7 +28,7 @@ import org.pjsip.pjsua2.*;
 interface MyAppObserver
 {
     abstract void notifyRegState(pjsip_status_code code, String reason,
-				 int expiration);
+				 long expiration);
     abstract void notifyIncomingCall(MyCall call);
     abstract void notifyCallState(MyCall call);
     abstract void notifyCallMediaState(MyCall call);
